\name{FMSmirfpckg}
\docType{data}
\alias{FMSmirfpckg}
\title{Functional Muscle Size and Strength}
\description{Data example used here comes from "Functional SNPs Associated with Muscle Size and Strength study" (FAMuSS) (Thompson et. al.,2004, Clarkson et al. 2005).  The FAMuSS study was conducted to determine the effects of single nucleotide polymorphisms (SNPs) on skeletal muscle size and strength before and after exercise training.  For the purposes of illustration, we use a subset of data including 2 genes namely ACTN3 and  RESISTIN which were shown to be associated with muscle size and strength.  ACTN3 has 4 SNPs and RESISTIN has 6 SNPs. The trait under study is perecentage change in muscle strength and is labelled NDRM.CH in the dataset.  Covariates are Center, Age and Gender.  The covariate used for HWE grouping is Race, which has 5 categories. 
}

\usage{data(FMSmirfpckg)}
\format{A data frame containing 991 rows and 16 columns (the first column is ID, column 2 to column 15 are genotypes, followed by Center, Gender, Age, Race, and NDRM.CH) 
}

\source{Thompson et al.}
\references{
Functional polymorphisms associated with human muscle size and strength. 
Thompson, P.D., Moyna, N. Seip, R., Clarkson, P., Angelopoulos, T., Gordon, P.,Pescatello, L., Visich, P., Zoeller, R., Devaney, J.M., Gordish, H., Bilbie, S., Hoffman, E.P. Med Sci Sports Exerc.  2004; 36(7): 1132-9

The grant number for FAMuSS is R01NS40606-02-Hoffman. 
}

\keyword{datasets}

